refactor: simplify game state initialization and handling
- Removed redundant START_GAME action and consolidated logic into INIT_GAME.
- Moved game initialization logic to cardUtils and refactored context initialization.
- Updated component usage to reflect new INIT_GAME action and simplified startNewGame function.
- Streamlined imports and removed unused imports from GameContext.
