+++ src/client/components/layouts/GamePage/NewPage.tsx
@@ -3,22 +3,19 @@ import Button from '../../elements/Button/Button'
 import { GameContext } from '../../../../api/GameContext'
 import { GameState } from '../../../../api/GameState'
 import { useContext } from 'react'
+import { startNewGame } from '../../../../api/cardUtils'
+import { CardContext } from '../../../../api/CardContext'
 
 export default function NewPage() {
   const { state: gameState, dispatch: gameDispatch } = useContext(GameContext)
-
-  const startNewGame = async () => {
-    gameDispatch({
-      type: 'START_GAME',
-    })
-  }
+  const { state: cardState } = useContext(CardContext)
 
   return (
     <div className="new-page-container">
       <Button
         label="Start New Game"
         className="new-page-button"
-        onClick={() => startNewGame()}
+        onClick={() => startNewGame(cardState.cards, gameDispatch)}
       />
       <p className="new-page-label">
         {gameState.gameState === GameState.ONGOING
